projects/Financial Planning/cost_allocation/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets integration
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.95.0

# OpenAI integration
openai>=1.0.0

# Data processing and analysis
python-dateutil>=2.8.2

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
# unittest-mock>=1.0.0

# Development and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: For enhanced data analysis
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For web interface (future enhancement)
flask>=2.3.0
flask-cors>=4.0.0

# Optional: For async operations (future enhancement)
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0 